requirements.txt
streamlit
Pillow
pytesseract
langchain
google-generativeai
python-dotenv
pdf2image 